:recycle: REFACTOR: converted useReports hook to fetchReports function
